requirements.txt
# Core dependencies
python-dotenv==1.0.0
pillow==10.0.0
pytesseract==0.3.10
numpy==1.24.3
requests==2.31.0
google-generativeai==0.3.1

# Async utilities
asyncio==3.4.3
aiohttp==3.8.5
aiodns==3.0.0

# Optional: Database connections
asyncpg==0.28.0
psycopg2-binary==2.9.7

# Optional: For embedding models
transformers==4.30.2
torch==2.0.1
sentence-transformers==2.2.2

# API and Web Framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
starlette>=0.26.1

# Database
pgvector>=0.1.8

# Integrations
pyairtable>=1.4.0
google-api-python-client>=2.70.0
google-auth>=2.17.3
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Utilities
requests==2.31.0
Pillow==10.0.0
loguru>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Embeddings & AI
voyageai==0.9.8
# Add LLM client libraries as needed
# openai>=1.0.0 # Removed
google-generativeai>=0.5.0 # Added Google Gemini client

# Text Processing (for chunking)
langchain-text-splitters>=0.0.1
tqdm==4.66.1

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

# New additions
aiohttp==3.8.5
asyncio==3.4.3 